+++ src/index.js
@@ -1,18 +1,6 @@
 import "./style.css";
 import tvLogo from "../images/tvlogo.png";
-import displayModal from "./displayModal.js";
 import displayScreen from "./modules/display.js";
-
-const list = document.querySelector(".shows");
-
-const getShows = async () => {
-  const response = await fetch("https://api.tvmaze.com/shows");
-  const JsonResponse = await response.json();
-  const mostRatedShows = JsonResponse.filter(
-    (show) => show.rating.average > 8.5
-  );
-  return mostRatedShows;
-};
 
 const navLogoContainer = document.querySelector(".logo");
 const imageLogo = document.createElement("img");
@@ -21,5 +9,3 @@ imageLogo.src = tvLogo;
 navLogoContainer.appendChild(imageLogo);
 
 displayScreen();
-
-export { getShows, list };
