+++ src/modules/getSingleShow.js
